+++ src/views/components/l13-diff-input/l13-diff-input.component.ts
@@ -53,6 +53,8 @@ export class L13DiffInputComponent extends L13Element<L13DiffInputViewModel> {
 		//	Update of menu is async and should be done right after. Fixes it.
 			await menu.viewmodel.update();
 			
+			menu.setFirstItemFocus();
+			
 		});
 		
 		this.input.addEventListener('click', async () => {
@@ -66,6 +68,8 @@ export class L13DiffInputComponent extends L13Element<L13DiffInputViewModel> {
 		//	Weird focus blur switch if click is outside of panel view.
 		//	Update of menu is async and should be done right after. Fixes it.
 			await menu.viewmodel.update();
+			
+			menu.setFirstItemFocus();
 			
 		});